requirements.txt
clipspy==1.0.0
